Crate trillium_compression

Body compression for trillium.rs

Currently, this crate only supports compressing outbound bodies with the zstd, brotli, and gzip algorithms (in order of preference), although more algorithms may be added in the future. The correct algorithm will be selected based on the Accept-Encoding header sent by the client, if one exists.

Structs

Compression

Trillium handler for compression

Enums

CompressionAlgorithm

Algorithms supported by this crate

Functions

compression

Alias for Compression::new